Move product disabler to navbar
